requirements.txt
flask==2.3.3
flask-socketio==5.3.4
pywebview==4.3
simple-websocket==1.0.0
Werkzeug==2.3.7
pyinstaller==6.12.0
schedule==1.2.0
tqdm==4.66.1
python-dotenv==1.0.0
pydantic==1.10.8
psutil==5.9.0
pytest==7.3.1
pytest-cov==4.1.0
